health commands now working
